[HACK] Work around name clash in Rails 7
As noted in
https://github.com/attr-encrypted/attr_encrypted/issues/379#issuecomment-818767027,
in Rails 7 uses a class method named `encrypted_attributes`, which
clashes with the `encrypted_attributes` method used in `attr_encrypted`.
As a result, when running the attr_encrypted test suite with Rails 7,
the test fails with the following error:

> lib/attr_encrypted.rb:176:in `block in attr_encrypted': undefined method `[]=' for nil:NilClass (NoMethodError)
>
> rake aborted!

To resolve the naming clash that causes this issue, this commit renames
the class method used in `attr_encrypted` from `encrypted_attributes` to
`third_party_encrypted_attributes` (which is admittedly verbose 🙈).
This change seems to get the test suite passing again on Rails 7.

Note: I see two test failures when running the attr_encrypted test suite
with Rails 7, but I see these same two failures when running the test
suite with Rails 6.1.4:

>   1) Failure:
> ActiveRecordTest#test_attribute_was_works_when_options_for_old_encrypted_value_are_different_than_options_for_new_encrypted_value [test/active_record_test.rb:217]:
> Expected: "password"
>   Actual: nil
>
>   2) Failure:
> ActiveRecordTest#test_should_create_was_predicate [test/active_record_test.rb:196]:
> Expected: "test@example.com"
>   Actual: nil